Move gamepad items to its own directory
