+++ src/gamepad/standard.ts
@@ -1,0 +1,26 @@
+export enum StandardButtons {
+	a = 0,
+	b = 1,
+	x = 2,
+	y = 3,
+	l1 = 4,
+	r1 = 5,
+	l2 = 6,
+	r2 = 7,
+	select = 8,
+	start = 9,
+	l3 = 10,
+	r3 = 11,
+	up = 12,
+	down = 13,
+	left = 14,
+	right = 15,
+	_unused = 16,
+}
+
+export enum StandardAxes {
+	leftX = 0,
+	leftY = 1,
+	rightX = 2,
+	rightY = 3,
+}
