+++ src/page-state/DebugPageState.tsx
@@ -5,9 +5,9 @@ import {AllowScroll} from '../AllowScroll.js';
 import {CircularButton} from '../circular-button.js';
 import {NavColumn} from '../navigation/NavColumn.js';
 import {Title} from '../Title.js';
-import {DebugGamepad} from './DebugGamepad.js';
+import {DebugGamepad} from '../gamepad/DebugGamepad.js';
+import {GamepadToSignal} from '../gamepad/GamepadToSignal.js';
 import {DebugTitle} from './DebugTitle.js';
-import {GamepadToSignal} from './GamepadToSignal.js';
 
 export const DebugPageState: FunctionComponent = () => (
 	<GamepadToSignal>
