+++ src/page-state/button.tsx
@@ -1,6 +1,6 @@
 import type {RefObject} from 'preact';
 import * as gfbi from '../icons/gamepad/face-button/index.js';
-import {StandardButtons} from './gamepad.js';
+import {StandardButtons} from '../gamepad/standard.js';
 import type {PageStateHooks, PageStateNode} from './node.js';
 
 const gamepadButtonHeld = new WeakSet<PageStateNode>();
